Update: package naming conventions. references - http://bit.ly/nameing-util-classes - http://bit.ly/naming-packages-java
