+++ src/main/java/com/imagelab/component/Draggable.java
@@ -1,0 +1,16 @@
+package com.imagelab.component;
+
+import javafx.scene.input.MouseEvent;
+
+/**
+ * Interface which contains the methods related to the
+ * draggable logic.
+ */
+public interface Draggable {
+    /**
+     * To detect the dragEvent.
+     *
+     * @param event - MouseEvent.
+     */
+    void dragDetected(MouseEvent event);
+}
