--- src/main/java/com/imagelab/components/Draggable.java
@@ -1,16 +1,0 @@
-package com.imagelab.components;
-
-import javafx.scene.input.MouseEvent;
-
-/**
- * Interface which contains the methods related to the
- * draggable logic.
- */
-public interface Draggable {
-    /**
-     * To detect the dragEvent.
-     *
-     * @param event - MouseEvent.
-     */
-    void dragDetected(MouseEvent event);
-}
