--- src/main/java/com/imagelab/operators/imagebluring/ApplyMedianBlurEffect.java
@@ -1,124 +1,0 @@
-package com.imagelab.operators.imagebluring;
-
-import com.imagelab.operators.OpenCVOperator;
-import com.imagelab.operators.basic.ReadImage;
-import com.imagelab.operators.basic.WriteImage;
-import com.imagelab.operators.geotransformation.RotateImage;
-import com.imagelab.operators.imageconversion.ColoredImageToBinary;
-import com.imagelab.operators.imageconversion.ConvertToGrayscale;
-import org.opencv.core.Mat;
-import org.opencv.imgproc.Imgproc;
-
-import java.util.HashSet;
-import java.util.Set;
-
-/**
- * Operator class which contains the logic related to the
- * applying median blur effects to an image  UI element.
- */
-public class ApplyMedianBlurEffect extends OpenCVOperator {
-    /**
-     * Information related to the ApplyMedianBlurEffect operator.
-     */
-    public enum Information {
-        /**
-         * Operator information in string format.
-         */
-        OPERATOR_INFO {
-            /**
-             * @return - Operator information and name of the operator.
-             */
-            public String toString() {
-                return "Apply median blur\n\nThis operations allows"
-                        + " you to apply median blur effects to your image."
-                        + " Moreover, you can change kernel value from"
-                        + " the properties.";
-            }
-        }
-    }
-
-    // properties of median blur effect.
-    private int kernelSize = 15;
-
-    /**
-     * This method contains the logic which validates the applicable
-     * openCV operations for a particular openCV operator.
-     *
-     * @param previous - accepts the previous operator to validate.
-     * @return - whether the received operator is valid or not.
-     */
-    @Override
-    public boolean validate(OpenCVOperator previous) {
-        if (previous == null) {
-            return false;
-        }
-        return allowedOperators().contains(previous.getClass());
-    }
-
-    /**
-     * This method contains the openCV operator related specific logic.
-     *
-     * @param image - accepts the mat object processed from the previous steps.
-     * @return - processed computed Mat obj.
-     */
-    @Override
-    public Mat compute(Mat image) {
-        return applyMedianBlurEffect(image, getKernelSize());
-    }
-
-    /**
-     * This method contains the applicable openCV operators for the selected
-     * openCV operator.
-     *
-     * @return - applicable operators.
-     */
-    @Override
-    public Set<Class<?>> allowedOperators() {
-        Set<Class<?>> allowed = new HashSet<>();
-        allowed.add(ReadImage.class);
-        allowed.add(RotateImage.class);
-        allowed.add(WriteImage.class);
-        allowed.add(ColoredImageToBinary.class);
-        allowed.add(ConvertToGrayscale.class);
-        allowed.add(ApplyBlurEffect.class);
-        allowed.add(ApplyGaussianBlurEffect.class);
-        allowed.add(ApplyMedianBlurEffect.class);
-        return allowed;
-    }
-
-    /**
-     * This method contains applying median blur related
-     * opencv logic.
-     *
-     * @param imageFile  - source file.
-     * @param kernelSize - size of the kernel. Should be int odd value.
-     * @return - median blur applied mat obj.
-     */
-    private Mat applyMedianBlurEffect(Mat imageFile, int kernelSize) {
-        // Creating an empty matrix to store the result.
-        Mat image = new Mat();
-
-        // Applying MedianBlur on the Image.
-        Imgproc.medianBlur(imageFile, image, kernelSize);
-        return image;
-    }
-
-    /**
-     * To get the kernel size related
-     * to the blur effect.
-     *
-     * @return kernelSize.
-     */
-    public int getKernelSize() {
-        return kernelSize;
-    }
-
-    /***
-     * To set the kernel size related
-     * to the blur effect.
-     * @param kernelSize - int.
-     */
-    public void setKernelSize(int kernelSize) {
-        this.kernelSize = kernelSize;
-    }
-}
