+++ src/main/java/com/imagelab/view/AbstractInformationUI.java
@@ -1,0 +1,10 @@
+package com.imagelab.view;
+
+import javafx.scene.layout.AnchorPane;
+
+/**
+ * Abstract class of information UI.
+ */
+public abstract class AbstractInformationUI extends AnchorPane {
+    //No implementation required.
+}
