+++ src/main/java/com/imagelab/view/forms/AbstractPropertiesFormUI.java
@@ -1,0 +1,12 @@
+package com.imagelab.view.forms;
+
+import javafx.scene.layout.AnchorPane;
+
+
+/**
+ * Abstract class which of properties UI forms and this extends from
+ * Anchor pane.
+ */
+public abstract class AbstractPropertiesFormUI extends AnchorPane {
+    //No implementation required.
+}
