--- src/main/java/com/imagelab/views/AbstractInformationUI.java
@@ -1,10 +1,0 @@
-package com.imagelab.views;
-
-import javafx.scene.layout.AnchorPane;
-
-/**
- * Abstract class of information UI.
- */
-public abstract class AbstractInformationUI extends AnchorPane {
-    //No implementation required.
-}
