--- src/main/java/com/imagelab/views/forms/AbstractPropertiesFormUI.java
@@ -1,12 +1,0 @@
-package com.imagelab.views.forms;
-
-import javafx.scene.layout.AnchorPane;
-
-
-/**
- * Abstract class which of properties UI forms and this extends from
- * Anchor pane.
- */
-public abstract class AbstractPropertiesFormUI extends AnchorPane {
-    //No implementation required.
-}
